Fixing Hidden Merge Issues
So there were a few things that got merged without my knowledge. The
important thing to note here is that in order for each page of the main
screen to scroll and swipe, we must have the wrapper id and scroller
class assigned to each. We might consider switching the wrapper to a
class rather than an id just for good measure.