Playing with Hover for FAQs - No Big Updates
I forgot I was messing with the Hover effect for the FAQs. My intent was
to have the questions change color while the question was expanded, but
I wasn't quite able to figure that out. I also added spacing at the
bottom of the FAQ to accommodate the expansion of the last question,
otherwise it expands outside of the viewing area.