FINALLY
After a long overdue overhaul, the app's interface is working correctly.
Note that there are a bunch of changes, including the removal of many
files. I have tested this extensively and can confirm that the 2-Way
Scroll is working, and the layout looks pretty satisfying as well. I may
make a few minor tweaks as we proceed, but overall, I think this is how
the final product should function. :+1: