Style and Surveys
I was able to make at least a good start on the Survey page, which we
can discuss and add to if necessary. I also made a few changes to the
style of the app, including the background color as Tarsha requested (I
think that was how she wanted it) as well as a slight change to the
Navigation bar which I think looks smashing.